初级撒花~
功能基本完成，没啥大BUG了，明天继续测试

+++ seckill/src/main/java/com/nwafu/seckill/dto/SeckillExecution.java
@@ -11,35 +11,36 @@ import com.nwafu.seckill.enums.SeckillStatEnum;
  */
 public class SeckillExecution {
 
-    private Long seckillId;
+    private int seckillId;
 
     //秒杀执行结果状态
     private int state;
 
     //状态表示
+    //stateInfo表示state状态数字的中文解释，比如：秒杀成功、秒杀结束、秒杀系统异常等信息。
     private String stateInfo;
 
     //秒杀成功的订单对象
     private SeckillOrder seckillOrder;
 
-    public SeckillExecution(Long seckillId, SeckillStatEnum seckillStatEnum, SeckillOrder seckillOrder) {
+    public SeckillExecution(int seckillId, SeckillStatEnum seckillStatEnum, SeckillOrder seckillOrder) {
         this.seckillId = seckillId;
         this.state = seckillStatEnum.getState();
         this.stateInfo = seckillStatEnum.getStateInfo();
         this.seckillOrder = seckillOrder;
     }
 
-    public SeckillExecution(Long seckillId, SeckillStatEnum seckillStatEnum) {
+    public SeckillExecution(int seckillId, SeckillStatEnum seckillStatEnum) {
         this.seckillId = seckillId;
         this.state = seckillStatEnum.getState();
         this.stateInfo = seckillStatEnum.getStateInfo();
     }
 
-    public Long getSeckillId() {
+    public int getSeckillId() {
         return seckillId;
     }
 
-    public void setSeckillId(Long seckillId) {
+    public void setSeckillId(int seckillId) {
         this.seckillId = seckillId;
     }
